+++ src/de.rs
@@ -149,7 +149,7 @@ impl Config {
     fn _load_file(path: &Path) -> Result<Self> {
         let buf = fs::read_to_string(path)
             .with_context(|| format!("failed to read `{}`", path.display()))?;
-        let mut config: Config = toml::from_str(&buf).with_context(|| {
+        let mut config: Config = basic_toml::from_str(&buf).with_context(|| {
             format!("failed to parse `{}` as cargo configuration", path.display())
         })?;
         config.set_path(path);