+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.1.9] - 2023-08-22
+
 - Recognize unstable `target.cfg(relocation_model = "...")` on nightly.
 
 ## [0.1.8] - 2023-07-03
@@ -54,7 +56,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/cargo-config2/compare/v0.1.8...HEAD
+[Unreleased]: https://github.com/taiki-e/cargo-config2/compare/v0.1.9...HEAD
+[0.1.9]: https://github.com/taiki-e/cargo-config2/compare/v0.1.8...v0.1.9
 [0.1.8]: https://github.com/taiki-e/cargo-config2/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/taiki-e/cargo-config2/compare/v0.1.6...v0.1.7
 [0.1.6]: https://github.com/taiki-e/cargo-config2/compare/v0.1.5...v0.1.6